Added core gameplay.
Still to do : add the cards to the memo of the winner, and give back cards to original ower if tie
Bonus : add the true rule when tie : draw one card without showing it for each player, draw another on top, visible face this time. Then compare cards and the winner wins them all.